+++ p2panda-rs/src/wasm/tests/entry.rs
@@ -4,9 +4,9 @@ use std::convert::TryFrom;
 
 use wasm_bindgen_test::*;
 
-use crate::document::DocumentId;
+use crate::document::DocumentViewId;
 use crate::hash::Hash;
-use crate::operation::{OperationEncoded, OperationFields, OperationValue, Relation};
+use crate::operation::{OperationEncoded, OperationFields, OperationValue, PinnedRelation};
 use crate::schema::SchemaId;
 use crate::test_utils::utils::create_operation;
 use crate::wasm::{decode_entry, sign_encode_entry, KeyPair, SignEncodeEntryResult};
@@ -16,9 +16,9 @@ wasm_bindgen_test_configure!(run_in_browser);
 #[wasm_bindgen_test]
 fn encodes_decodes_entries() {
     let key_pair = KeyPair::new();
-    let schema = SchemaId::Application(Relation::new(DocumentId::new(
+    let schema = SchemaId::Application(PinnedRelation::new(DocumentViewId::new(vec![
         Hash::new_from_bytes(vec![1, 2, 3]).unwrap(),
-    )));
+    ])));
 
     let mut fields = OperationFields::new();
     fields